feat(home) comments creation

+++ routes/home.js
@@ -26,11 +26,13 @@ router.put('/post', async function (req, res) {
             if (idPost == null)
                 post = await Post.find()
                     .populate('user')
+                    .populate('categories')
                     .sort({'_id': -1})
                     .limit(10);
             else
                 post = await Post.find()
                     .populate('user')
+                    .populate('categories')
                     .where('_id').lt(idPost)
                     .sort({'_id': -1})
                     .limit(10);
@@ -140,55 +142,69 @@ router.put('/groups', async function(req, res) {
 });
 
 router.post('/post/create', async function (req, res) {
-    const {id, title, description, categories} = req.body;
+    let {id, title, description, categories} = req.body;
     const files = req.files;
     let images = [];
     const date = new Date();
+    if(!Array.isArray(categories))
+        categories = [categories];
 
-    const newPost = new Post({"user": id, "categories":categories, "date":date,"title": title, "post": description});
+    const newPost = new Post({"user": id, "categories": categories, "date": date, "title": title, "post": description});
 
     const post = await newPost.save().catch(err => console.log(err));
-    for (const file of files) {
-        const {path} = file;
-        const newPath = await cloudinary.upload(path, 'shareart/users/' + id + '/posts/' + post._id, null)
-        images.push(newPath)
-        fs.removeSync(path);
+    if (files !== undefined) {
+        for (const file of files) {
+            const {path} = file;
+            const newPath = await cloudinary.upload(path, 'shareart/users/' + id + '/posts/' + post._id, null)
+            images.push(newPath)
+            fs.removeSync(path);
+        }
+        post.images = images;
     }
-    post.images = images;
-
     await post.save().then(response => res.json(response));
 });
 
-router.put('/post/edit', async function (req, res){
-    const {id, title, description, categories, idImages} = req.body;
+router.put('/post/edit', async function (req, res) {
+    const {id, title, description, categories, deleteImages} = req.body;
     const files = req.files;
 
     const post = await Post.findOne({'_id': id});
 
     post.title = title;
     post.description = description;
-    post.categories = categories;
+    if(Array.isArray(categories))
+        post.categories = categories;
+    else
+        post.categories = [categories];
 
-    idImages.forEach(id => {
+    if (Array.isArray(deleteImages)) {
+        deleteImages.forEach(id => {
+            for (let i = 0; i < post.images.length; i++) {
+                if (post.images[i].id === id) {
+                    post.images.splice(i, 1);
+                    i--;
+                }
+            }
+        });
+        deleteImages.forEach(image => cloudinary.delete(image));
+    } else {
         for (let i = 0; i < post.images.length; i++) {
-            if (post.images[i].id === id) {
+            if (post.images[i].id === deleteImages) {
                 post.images.splice(i, 1);
                 i--;
             }
         }
-    });
-
-    idImages.forEach(image => cloudinary.delete(image));
-
-    for (const file of files) {
-        const {path} = file;
-        const newPath = await cloudinary.upload(path, 'shareart/users/' + post.user + '/posts/' + post._id, null)
-        post.images.push(newPath)
-        fs.removeSync(path);
+        cloudinary.delete(deleteImages);
     }
 
-    post.save().then(response => res.json(response))
+    files?.forEach(file => {
+        const {path} = file;
+        const newPath = cloudinary.upload(path, 'shareart/users/' + post.user + '/posts/' + post._id, null)
+        post.images.push(newPath)
+        fs.removeSync(path);
+    })
 
+    post.save().then(response => res.json(response))
 });
 
 router.post('/post/delete', async function (req, res) {
@@ -208,10 +224,47 @@ router.post('/post/delete', async function (req, res) {
 
 router.put('/post/vote', async function (req, res) {
     const {idUser, type, idPost} = req.body;
-    const post = Post.find({})
+    const post = await Post.find({'_id': idPost})
+
+    for (let i = 0; i < post.votes.length; i++) {
+        if (post.votes[i].idUser === idUser) {
+            post.votes.splice(i, 1);
+            break;
+        }
+    }
+
+    if (type !== 2) {
+        const vote = {"idUser": idUser, "action": type};
+        post.votes.push(vote);
+    }
+
+    await post.save().then(response => res.json(response));
+
 });
 
-router.put('/comment/create', async function (req, res){
+router.put('/comment/create', async function (req, res) {
+    const {idUser, idPost, idComment, content} = req.body;
+    const date = new Date();
+    const newComment = new Comment({"user": idUser,"post": idPost , "comment": content, "date": date});
+    const comment = await newComment.save();
+
+    if(idComment != null){
+        const fatherComment = await Comment.find({'_id':idComment});
+        fatherComment.comments.push(comment._id);
+        await fatherComment.save();
+    }
+
+    res.status(200);
+    res.json({"message":"ok"})
+});
+
+router.post('/comment/edit', async function (req, res){
+    const {id,content} = req.body;
+    const comment = await Comment.find({'_id':id});
+
+    comment.comment = content;
+    comment.date
+
 
 
 });
@@ -221,8 +274,23 @@ router.post('/comment/delete', async function (req, res){
 
 });
 
-router.put('/comment/vote', async function (req, res){
+router.put('/comment/vote', async function (req, res) {
+    const {idUser, type, idComment} = req.body;
+    const comment = await Comment.find({'_id': idComment})
 
+    for (let i = 0; i < comment.votes.length; i++) {
+        if (comment.votes[i].idUser === idUser) {
+            comment.votes.splice(i, 1);
+            break;
+        }
+    }
+
+    if (type !== 2) {
+        const vote = {"idUser": idUser, "action": type};
+        comment.votes.push(vote);
+    }
+
+    await comment.save().then(response => res.json(response));
 
 });
